docs: checkNumberStatus
