+++ src/lib/wapi/functions/get-profile-pic-from-server.js
@@ -52,6 +52,7 @@ MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNMMNMNMMMNMMNNMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMNNNNMMNNNMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 MMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMMM
 */
-export function getProfilePicFromServer(id) {
-  return Store.WapQuery.profilePicFind(id).then((x) => x.eurl);
+export async function getProfilePicFromServer(id) {
+  const pinc = await Store.WapQuery.profilePicFind(id).then((x) => x.eurl);
+  return pinc;
 }
