+++ src/lib/wapi/serializers/serialize-number-status.js
@@ -63,7 +63,8 @@ export const _serializeNumberStatusObj = (obj) => {
       id: obj.jid,
       status: obj.status,
       isBusiness: obj.biz === true,
-      canReceiveMessage: obj.status === 200
+      canReceiveMessage: obj.status === 200,
+      profilePic: undefined
     }
   );
 };
