+++ src/lib/wapi/wapi.js
@@ -162,7 +162,8 @@ import {
   setGroupSettings,
   sendButtons,
   sendListMenu,
-  checkChat
+  checkChat,
+  checkNumberStatus
 } from './functions';
 import {
   base64ToFile,
@@ -238,6 +239,7 @@ if (typeof window.WAPI === 'undefined') {
   window.WAPI.returnReply = returnReply;
   window.WAPI.getStore = getStore;
   window.WAPI.checkChat = checkChat;
+  window.WAPI.checkNumberStatus = checkNumberStatus;
 
   //Profile
   window.WAPI.setProfilePic = setProfilePic;
@@ -457,38 +459,6 @@ if (typeof window.WAPI === 'undefined') {
         ''
       )
     );
-  };
-
-  window.WAPI._serializeNumberStatusObj = (obj) => {
-    if (obj == undefined) {
-      return null;
-    }
-
-    return Object.assign(
-      {},
-      {
-        id: obj.jid,
-        status: obj.status,
-        isBusiness: obj.biz === true,
-        canReceiveMessage: obj.status === 200
-      }
-    );
-  };
-
-  window.WAPI.checkNumberStatus = async function (id) {
-    try {
-      const result = await window.Store.WapQuery.queryExist(id);
-      if (result.status === 404) throw 404;
-      if (result.jid === undefined) throw 404;
-      const data = window.WAPI._serializeNumberStatusObj(result);
-      if (data.status == 200) data.numberExists = true;
-      return data;
-    } catch (e) {
-      return window.WAPI._serializeNumberStatusObj({
-        status: e,
-        jid: new window.Store.WidFactory.createWid(id)
-      });
-    }
   };
 
   window.WAPI.getChatIsOnline = async function (chatId) {
